internal/worker: compute and display stats
Keep track of load-shedding statistics.

Display them on the worker home page, along
with various memory statistics that help
to understand the worker's memory usage.

Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/255977
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>

+++ internal/worker/fetch.go
@@ -176,7 +176,7 @@ func fetchAndInsertModule(ctx context.Context, modulePath, requestedVersion stri
 	start := time.Now()
 	minfo := fetch.GetModuleInfo(ctx, modulePath, requestedVersion, proxyClient)
 	if minfo.Error == nil {
-		shouldShed, deferFunc := decideToShed(minfo.ZipSize)
+		shouldShed, deferFunc := decideToShed(uint64(minfo.ZipSize))
 		defer deferFunc()
 		if shouldShed {
 			ft.Error = derrors.SheddingLoad
